task-executor/api/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Async support
asyncpg==0.29.0
greenlet==3.0.3

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development
black==24.1.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
click==8.1.7